# docs/generated/services/TenantService/openapi.yml
openapi: 3.0.4
info:
  version: 0.3.0
  title: Tenant
paths:
...
  /api/Facility/{facilityId}:
    get:
      tags:
      - Facility
      summary: Find a facility config by Id
      parameters:
      - name: facilityId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/FacilityConfig'
            application/json:
              schema:
                $ref: '#/components/schemas/FacilityConfig'
            text/json:
              schema:
                $ref: '#/components/schemas/FacilityConfig'
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "500":
          description: Internal Server Error
    delete:
      tags:
      - Facility
      summary: Delete a facility by Id.
      parameters:
      - name: facilityId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
...
  /api/Facility/{id}:
    put:
      tags:
      - Facility
      summary: Update a facility config.
      parameters:
      - name: id
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacilityConfig'
          text/json:
            schema:
              $ref: '#/components/schemas/FacilityConfig'
          application/*+json:
            schema:
              $ref: '#/components/schemas/FacilityConfig'
      responses:
        "201":
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/FacilityConfig'
            application/json:
              schema:
                $ref: '#/components/schemas/FacilityConfig'
            text/json:
              schema:
                $ref: '#/components/schemas/FacilityConfig'
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "500":
          description: Internal Server Error
...
components:
  schemas:
...
    FacilityConfig:
      type: object
      properties:
        id:
          type: string
          nullable: true
        facilityId:
          type: string
          nullable: true
        facilityName:
          type: string
          nullable: true
        timeZone:
          type: string
          nullable: true
        isDeleted:
          type: boolean
          default: false
        scheduledReports:
          $ref: '#/components/schemas/TenantScheduledReportConfig'
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: {}
...
    TenantScheduledReportConfig:
      type: object
      properties:
        daily:
          type: array
          items:
            type: string
          nullable: true
        weekly:
          type: array
          items:
            type: string
          nullable: true
        monthly:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false